+++ projects/hipcub/hipcub/include/hipcub/backend/rocprim/warp/warp_reduce.hpp
@@ -38,11 +38,33 @@
 
 BEGIN_HIPCUB_NAMESPACE
 
-template<typename T, int LOGICAL_WARP_THREADS = HIPCUB_DEVICE_WARP_THREADS>
-class WarpReduce : private ::rocprim::warp_reduce<T, LOGICAL_WARP_THREADS>
+namespace detail
 {
-    static_assert(LOGICAL_WARP_THREADS > 0, "LOGICAL_WARP_THREADS must be greater than 0");
-    using base_type = typename ::rocprim::warp_reduce<T, LOGICAL_WARP_THREADS>;
+constexpr auto sum_reduce_op = _HIPCUB_STD::plus<>{};
+
+constexpr auto max_reduce_op =
+#if _HIPCUB_HAS_DEVICE_SYSTEM_STD
+    _HIPCUB_LIBCXX::maximum<>{}
+#else
+    [](auto a, auto b) { return a > b ? a : b; }
+#endif
+;
+
+constexpr auto min_reduce_op =
+#if _HIPCUB_HAS_DEVICE_SYSTEM_STD
+    _HIPCUB_LIBCXX::minimum<>{}
+#else
+    [](auto a, auto b) { return a > b ? b : a; }
+#endif
+;
+} // namespace detail
+
+template<typename T, int LogicalWarpThreads = HIPCUB_DEVICE_WARP_THREADS>
+class WarpReduce : private ::rocprim::warp_reduce<T, LogicalWarpThreads>
+{
+    static_assert(LogicalWarpThreads >= 1 && LogicalWarpThreads <= HIPCUB_DEVICE_WARP_THREADS,
+                  "LogicalWarpThreads must be in the range [1, HIPCUB_DEVICE_WARP_THREADS]");
+    using base_type = typename ::rocprim::warp_reduce<T, LogicalWarpThreads>;
 
     typename base_type::storage_type &temp_storage_;
 
@@ -61,10 +83,68 @@ public:
         return input;
     }
 
-    HIPCUB_DEVICE inline
-    T Sum(T input, int valid_items)
+    template<typename InputT,
+             HIPCUB_REQUIRES(HIPCUB_TRAIT(detail::is_fixed_size_random_access_range, InputT))>
+    HIPCUB_DEVICE
+    inline T Sum(const InputT& input)
+    {
+        auto thread_reduction = HIPCUB_NAMESPACE::ThreadReduce(input, detail::sum_reduce_op);
+        base_type::reduce(thread_reduction, thread_reduction, temp_storage_);
+        return thread_reduction;
+    }
+
+    HIPCUB_DEVICE
+    inline T Sum(T input, int valid_items)
     {
         base_type::reduce(input, input, valid_items, temp_storage_);
+        return input;
+    }
+
+    HIPCUB_DEVICE
+    inline T Max(T input)
+    {
+        base_type::reduce(input, input, temp_storage_, detail::max_reduce_op);
+        return input;
+    }
+
+    template<typename InputT,
+             HIPCUB_REQUIRES(HIPCUB_TRAIT(detail::is_fixed_size_random_access_range, InputT))>
+    HIPCUB_DEVICE
+    inline T Max(const InputT& input)
+    {
+        auto thread_reduction = HIPCUB_NAMESPACE::ThreadReduce(input, detail::max_reduce_op);
+        base_type::reduce(thread_reduction, thread_reduction, temp_storage_, detail::max_reduce_op);
+        return thread_reduction;
+    }
+
+    HIPCUB_DEVICE
+    inline T Max(T input, int valid_items)
+    {
+        base_type::reduce(input, input, valid_items, temp_storage_, detail::max_reduce_op);
+        return input;
+    }
+
+    HIPCUB_DEVICE
+    inline T Min(T input)
+    {
+        base_type::reduce(input, input, temp_storage_, detail::min_reduce_op);
+        return input;
+    }
+
+    template<typename InputT,
+             HIPCUB_REQUIRES(HIPCUB_TRAIT(detail::is_fixed_size_random_access_range, InputT))>
+    HIPCUB_DEVICE
+    inline T Min(const InputT& input)
+    {
+        auto thread_reduction = HIPCUB_NAMESPACE::ThreadReduce(input, detail::min_reduce_op);
+        base_type::reduce(thread_reduction, thread_reduction, temp_storage_, detail::min_reduce_op);
+        return thread_reduction;
+    }
+
+    HIPCUB_DEVICE
+    inline T Min(T input, int valid_items)
+    {
+        base_type::reduce(input, input, valid_items, temp_storage_, detail::min_reduce_op);
         return input;
     }
 
@@ -82,6 +162,17 @@ public:
     {
         base_type::tail_segmented_reduce(input, input, tail_flag, temp_storage_);
         return input;
+    }
+
+    template<typename InputT,
+             typename ReduceOp,
+             HIPCUB_REQUIRES(HIPCUB_TRAIT(detail::is_fixed_size_random_access_range, InputT))>
+    HIPCUB_DEVICE
+    inline T Reduce(const InputT& input, ReduceOp reduce_op)
+    {
+        auto thread_reduction = HIPCUB_NAMESPACE::ThreadReduce(input, reduce_op);
+        base_type::reduce(thread_reduction, thread_reduction, temp_storage_, reduce_op);
+        return thread_reduction;
     }
 
     template<typename ReduceOp>
@@ -120,6 +211,139 @@ public:
         return input;
     }
 };
+
+#ifndef DOXYGEN_SHOULD_SKIP_THIS // Do not document
+template<typename T>
+class WarpReduce<T, 1>
+{
+private:
+    using _TempStorage = hipcub::NullType;
+
+public:
+    struct TempStorage : Uninitialized<_TempStorage>
+    {};
+
+    HIPCUB_DEVICE
+    inline WarpReduce(TempStorage& /*temp_storage*/)
+    {}
+
+    HIPCUB_DEVICE
+    inline T Sum(T input)
+    {
+        return input;
+    }
+
+    template<typename InputT,
+             HIPCUB_REQUIRES(HIPCUB_TRAIT(detail::is_fixed_size_random_access_range, InputT))>
+    HIPCUB_DEVICE
+    inline T Sum(const InputT& input)
+    {
+        auto thread_reduction = HIPCUB_NAMESPACE::ThreadReduce(input, detail::sum_reduce_op);
+        return thread_reduction;
+    }
+
+    HIPCUB_DEVICE
+    inline T Sum(T input, int /*valid_items*/)
+    {
+        return input;
+    }
+
+    HIPCUB_DEVICE
+    inline T Max(T input)
+    {
+        return input;
+    }
+
+    template<typename InputT,
+             HIPCUB_REQUIRES(HIPCUB_TRAIT(detail::is_fixed_size_random_access_range, InputT))>
+    HIPCUB_DEVICE
+    inline T Max(const InputT& input)
+    {
+        auto thread_reduction = HIPCUB_NAMESPACE::ThreadReduce(input, detail::max_reduce_op);
+        return thread_reduction;
+    }
+
+    HIPCUB_DEVICE
+    inline T Max(T input, int /*valid_items*/)
+    {
+        return input;
+    }
+
+    HIPCUB_DEVICE
+    inline T Min(T input)
+    {
+        return input;
+    }
+
+    template<typename InputT,
+             HIPCUB_REQUIRES(HIPCUB_TRAIT(detail::is_fixed_size_random_access_range, InputT))>
+    HIPCUB_DEVICE
+    inline T Min(const InputT& input)
+    {
+        auto thread_reduction = HIPCUB_NAMESPACE::ThreadReduce(input, detail::min_reduce_op);
+        return thread_reduction;
+    }
+
+    HIPCUB_DEVICE
+    inline T Min(T input, int /*valid_items*/)
+    {
+        return input;
+    }
+
+    template<typename FlagT>
+    HIPCUB_DEVICE
+    inline T HeadSegmentedSum(T input, FlagT /*head_flag*/)
+    {
+        return input;
+    }
+
+    template<typename FlagT>
+    HIPCUB_DEVICE
+    inline T TailSegmentedSum(T input, FlagT /*tail_flag*/)
+    {
+        return input;
+    }
+
+    template<typename InputT,
+             typename ReduceOp,
+             HIPCUB_REQUIRES(HIPCUB_TRAIT(detail::is_fixed_size_random_access_range, InputT))>
+    HIPCUB_DEVICE
+    inline T Reduce(const InputT& input, ReduceOp reduce_op)
+    {
+        auto thread_reduction = HIPCUB_NAMESPACE::ThreadReduce(input, reduce_op);
+        return thread_reduction;
+    }
+
+    template<typename ReduceOp>
+    HIPCUB_DEVICE
+    inline T Reduce(T input, ReduceOp /*reduce_op*/)
+    {
+        return input;
+    }
+
+    template<typename ReduceOp>
+    HIPCUB_DEVICE
+    inline T Reduce(T input, ReduceOp /*reduce_op*/, int /*valid_items*/)
+    {
+        return input;
+    }
+
+    template<typename ReduceOp, typename FlagT>
+    HIPCUB_DEVICE
+    inline T HeadSegmentedReduce(T input, FlagT /*head_flag*/, ReduceOp /*reduce_op*/)
+    {
+        return input;
+    }
+
+    template<typename ReduceOp, typename FlagT>
+    HIPCUB_DEVICE
+    inline T TailSegmentedReduce(T input, FlagT /*tail_flag*/, ReduceOp /*reduce_op*/)
+    {
+        return input;
+    }
+};
+
+#endif
 
 END_HIPCUB_NAMESPACE